refactor(cockpit): extract _RollupItem into sibling module (#1354)
Sixth wedge of the cockpit_ui.py god-module split tracked by #1354.
Moves the inbox rollup sub-item widget (66 LOC) into
cockpit_inbox_rollup_item.py and re-exports it from cockpit_ui for
back-compat so existing call sites (and tests that import via
pollypm.cockpit_ui) continue to work unchanged.

cockpit_ui.py drops from 18,651 to 18,585 LOC.

Prior extractions:
- _InboxProjectPickerModal (#1606)
- _FirstShippedCelebrationModal (#1618)
- _SettingsConfirmModal (#1670)
- _AlertDetailModal (#1688)
- _SettingsAccountReassignModal (#1689)

_md_to_rich still lives in cockpit_ui, so the new module imports it
lazily inside _build_text to avoid an import cycle. The 4 rollup tests
in tests/test_cockpit_inbox_ui.py pass; full inbox UI suite (86 tests)
remains green.
